third_party/WebKit/LayoutTests/fast/dom/custom/prerender-insert-after-stop.html

<!DOCTYPE html>
<script src='../../../resources/js-test.js'></script>
<script>
description('Test that Custom Element should be defined while navigating between frames.');

var customBody;
var jsTestIsAsync = true;

function linkInserter(doc) {
    return function() {
        customBody = doc.getElementById('thebody');
        shouldBeDefined('customBody');
        shouldBeEqualToString('customBody.toLocaleString()', '[object HTMLBodyElement]');
        finishJSTest();
    }
}

function navigateIFrameThenInsertLink() {
    var iframe = document.getElementById('iframe');
    iframe.onload = linkInserter(iframe.contentDocument);
    iframe.contentWindow.location = 'resources/empty-custom-body.html';
}
</script>
<iframe id='iframe' onload='navigateIFrameThenInsertLink()' src='resources/empty-custom-body.html'>
</iframe>
