third_party/WebKit/LayoutTests/fast/dom/custom/prerender-insert-after-stop-expected.txt

Test that Custom Element should be defined while navigating between frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS customBody is defined.
PASS customBody.toLocaleString() is "[object HTMLBodyElement]"
PASS successfullyParsed is true

TEST COMPLETE
